updated some UI for customer

+++ madhan-customer/src/app/manage-pol-pod/pol-pod-list/pol-pod-list.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { PolPodApiService } from '../service/api/pol-pod-api.service';
 import { ActivatedRoute, Router } from '@angular/router';
+import { LoaderService } from 'src/app/service/Loader/loader.service';
+import { AlertServiceService } from 'src/app/service/alert/alert-service.service';
 
 @Component({
   selector: 'app-pol-pod-list',
@@ -10,10 +12,13 @@ import { ActivatedRoute, Router } from '@angular/router';
 export class PolPodListComponent implements OnInit {
 
   polPods = [];
-  constructor(private api: PolPodApiService, private router: Router) { }
+  constructor(private api: PolPodApiService,
+    private router: Router,
+    private loader: LoaderService,
+    private alert: AlertServiceService) { }
 
   ngOnInit() {
-    
+
   }
 
   ionViewWillEnter() {
@@ -21,10 +26,12 @@ export class PolPodListComponent implements OnInit {
   }
 
   getPolPods() {
+    this.loader.createLoader();
     this.api.getPolPods().subscribe((resp: any) => {
+      this.loader.dismissLoader();
       console.log(resp);
       this.polPods = resp || [];
-      
+
     }, err => {
 
     })
@@ -41,13 +48,20 @@ export class PolPodListComponent implements OnInit {
   }
 
   deletePolPod(item) {
-    console.log('delete = ', item);
-    this.api.deletePolPodById(item.polpodid).subscribe(res => {
-      console.log(res);
-      this.getPolPods();
-    }, err => {
-      console.log(err);
-      this.getPolPods();
-    })
+    this.alert.alertPromt().then(data => {
+      if (Boolean(data)) {
+        console.log('delete = ', item);
+        // this.loader.createLoader();
+        this.api.deletePolPodById(item.polpodid).subscribe(res => {
+          // this.loader.dismissLoader();
+          console.log(res);
+          this.getPolPods();
+        }, err => {
+          console.log(err);
+          this.getPolPods();
+        })
+      }
+    });
+    this.loader.dismissLoader();
   }
 }
